centralize args defaults
