+++ src/util.ts
@@ -1,18 +1,9 @@
-import * as fs from 'fs';
-import * as path from 'path';
 import * as debug from 'debug';
 import * as minimist from 'minimist';
 import { Args } from './model';
 import { valueOrThrow } from './scripts/webpack/util';
 
 debug.enable('scriptoni:*');
-
-type Key = Exclude<keyof Args, '_' | 'wiro' | 'bundleAnalyzer'>;
-
-export function loadFileFromArgument(args: Pick<Args, Key>, key: Key, defaultPath: string) {
-  const filePath = path.join(process.cwd(), args[key] || defaultPath);
-  return fs.existsSync(filePath) && require(filePath);
-}
 
 export const logger = {
   bin: debug('scriptoni:bin'),
@@ -24,6 +15,16 @@ export const logger = {
   prettier: debug('scriptoni:prettier')
 };
 
+const defaultArgs: Partial<Args> = {
+  paths: './paths.js',
+  metarpheusConfig: 'metarpheus-ts-config.js',
+  c: './config'
+};
+
 export const getArgs = (): Args => {
-  return valueOrThrow(Args, minimist(process.argv.slice(2)) as any);
+  const args = {
+    ...defaultArgs,
+    ...minimist(process.argv.slice(2))
+  };
+  return valueOrThrow(Args, args as any);
 };
